Add files via upload
The dictionaries in this part are made and mostly display properly. The only problem I haven't been able to fix so far is that the huffman codes don't display on the correct letter, but are all assigned correctly in the huffman dictionary itself. I left code in comments that will let you see the proper assignments in the console until I fix this one last problem.